reorged components and data
