+++ src/components/Button.js
@@ -1,0 +1,7 @@
+export const Button = ({ children, onClick, disabled }) => {
+	return (
+		<button className='button' onClick={onClick} disabled={disabled}>
+			{children}
+		</button>
+	);
+};
