+++ src/components/Friend.js
@@ -1,0 +1,24 @@
+import { Button } from './Button';
+
+export const Friend = ({ friend, onSelect, selected }) => {
+	const isSelected = selected?.id === friend.id;
+
+	return (
+		<li className={isSelected ? 'selected' : ''}>
+			<img src={friend.image} alt={`${friend.name}'s avatar`} />
+			<h3>{friend.name}</h3>
+			{friend.balance < 0 && (
+				<p className='red'>
+					You owe {friend.name} ${Math.abs(friend.balance)}
+				</p>
+			)}
+			{friend.balance > 0 && (
+				<p className='green'>
+					{friend.name} owes you ${friend.balance}
+				</p>
+			)}
+			{friend.balance === 0 && <p>You and {friend.name} are even</p>}
+			<Button onClick={() => onSelect(friend)}>{isSelected ? 'Close' : 'Select'}</Button>
+		</li>
+	);
+};
